Feature: Smart contract
### Why?

Every dApp needs a reliable smart contract to ensure that transactions
are secure and function as intended. This PR introduces the main project
contract along with tests, enabling seamless integration with the
frontend.

### Description

The contract tracks and stores the target word, the number of tries
left, all guesses, and the token used for each guess. It provides
functions to set the target word, try and register a guess, end the
game, and retrieve specific values from the contract, ensuring proper
tracking of gameplay and token management.