Feature: Contract integration
### Why?

With the core smart contract implemented, the next step is to connect it
to the webapp. This integration allows users to interact directly with
the blockchain through the frontend by enabling guess submissions,
result and tries tracking in a decentralized manner.

### Description

This PR connects the web application to the deployed smart contract
using a web3 provider (e.g., MetaMask). It includes functions for
initializing the contract instance, handling user wallet connections,
and executing contract calls such as submitting guesses, fetching game
data, and updating the UI based on blockchain responses.

### TODO

- [x] Get Random Word
- [x] Guess word input
- [x] Connect with the Contract
- [x] Hash Word and Guesses
- [x] Apply limited number of tries