Feat: Layout
### Why?

To fully play the game, this PR implements a interface inspired by
[Wordle](https://www.nytimes.com/games/wordle/index.html) and connects
with the Smart Contract functions.

### Description

It creates two main components: `Grid` and `Keyboard` that will provide
the track of letters and apply the classic color logic from the original
game. It also refactor some functions from the Contract that are not
being used anymore to make it lighter.